Move language file loading into own test file
